+++ typescript/cli/src/deploy/core.ts
@@ -12,7 +12,7 @@ import {
 } from '@hyperlane-xyz/sdk';
 
 import { MINIMUM_CORE_DEPLOY_GAS } from '../consts.js';
-import { getOrRequestApiKeys } from '../context/context.js';
+import { requestAndSaveApiKeys } from '../context/context.js';
 import { WriteCommandContext } from '../context/types.js';
 import { log, logBlue, logGray, logGreen } from '../logger.js';
 import { runSingleChainSelectionStep } from '../utils/chains.js';
@@ -34,6 +34,7 @@ interface DeployParams {
 interface ApplyParams extends DeployParams {
   deployedCoreAddresses: DeployedCoreAddresses;
 }
+
 /**
  * Executes the core deploy command.
  */
@@ -62,7 +63,7 @@ export async function runCoreDeploy(params: DeployParams) {
   }
   let apiKeys: ChainMap<string> = {};
   if (!skipConfirmation)
-    apiKeys = await getOrRequestApiKeys([chain], chainMetadata);
+    apiKeys = await requestAndSaveApiKeys([chain], chainMetadata, registry);
 
   const signer = multiProvider.getSigner(chain);