+++ typescript/cli/src/utils/input.ts
@@ -13,12 +13,18 @@ import {
   useState,
 } from '@inquirer/core';
 import figures from '@inquirer/figures';
-import { KeypressEvent, confirm, input } from '@inquirer/prompts';
+import { KeypressEvent, confirm, input, isSpaceKey } from '@inquirer/prompts';
 import type { PartialDeep } from '@inquirer/type';
 import ansiEscapes from 'ansi-escapes';
 import chalk from 'chalk';
 
-import { WarpCoreConfig } from '@hyperlane-xyz/sdk';
+import { ProxyAdmin__factory } from '@hyperlane-xyz/core';
+import {
+  ChainName,
+  DeployedOwnableConfig,
+  WarpCoreConfig,
+} from '@hyperlane-xyz/sdk';
+import { Address, isAddress, rootLogger } from '@hyperlane-xyz/utils';
 
 import { readWarpCoreConfig } from '../config/warp.js';
 import { CommandContext } from '../context/types.js';
@@ -75,6 +81,59 @@ export async function inputWithInfo({
     if (answer === INFO_COMMAND) logGray(indentedInfo);
   } while (answer === INFO_COMMAND);
   return answer;
+}
+
+/**
+ * Prompts the user to optionally set an existing ProxyAdmin contract address to be used in a WarpToken deployment.
+ */
+export async function setProxyAdminConfig(
+  context: CommandContext,
+  chain: ChainName,
+  warpRouteOwner: Address,
+): Promise<DeployedOwnableConfig> {
+  const defaultAdminConfig: DeployedOwnableConfig = {
+    owner: warpRouteOwner,
+  };
+
+  // default to deploying a new ProxyAdmin with `warpRouteOwner` as the owner
+  // if the user supplied the --yes flag
+  if (context.skipConfirmation) {
+    return defaultAdminConfig;
+  }
+
+  const useExistingProxy = await confirm({
+    message: `Use an existing Proxy Admin contract for the warp route deployment on chain "${chain}"?`,
+  });
+
+  if (!useExistingProxy) {
+    return defaultAdminConfig;
+  }
+
+  const proxyAdminAddress = await input({
+    message: `Please enter the address of the Proxy Admin contract to be used on chain "${chain}":`,
+    validate: isAddress,
+  });
+
+  const proxy = ProxyAdmin__factory.connect(
+    proxyAdminAddress,
+    context.multiProvider.getProvider(chain),
+  );
+
+  try {
+    const ownerAddress = await proxy.owner();
+    return {
+      address: proxyAdminAddress,
+      owner: ownerAddress,
+    };
+  } catch (error) {
+    rootLogger.error(
+      `Failed to read owner address from ProxyAdmin contract at ${proxy.address} on chain ${chain}.`,
+      error,
+    );
+    throw new Error(
+      `Failed to read owner address from ProxyAdmin contract at ${proxy.address}. Are you sure this is a ProxyAdmin contract?`,
+    );
+  }
 }
 
 /**
@@ -316,9 +375,9 @@ function getHelpTips({
   let helpTipBottom = '';
   const defaultTopHelpTip =
     instructions ??
-    `(Press ${theme.style.key('tab')} to select, and ${theme.style.key(
-      'enter',
-    )} to proceed`;
+    `(Press ${theme.style.key('tab')} or ${theme.style.key(
+      'space',
+    )} to select, and ${theme.style.key('enter')} to proceed`;
   const defaultBottomHelpTip = `\n${theme.style.help(
     '(Use arrow keys to reveal more choices)',
   )}`;
@@ -530,7 +589,10 @@ export const searchableCheckBox = createPrompt(
           );
           setActive(next);
         }
-      } else if (key.name === 'tab' && optionState.options.length > 0) {
+      } else if (
+        (key.name === 'tab' || isSpaceKey(key)) &&
+        optionState.options.length > 0
+      ) {
         // Avoid the message header to be printed again in the console
         rl.clearLine(0);