+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getEclipseStrideSTTIAWarpConfig.ts
@@ -6,20 +6,21 @@ import {
 } from '@hyperlane-xyz/sdk';
 
 export const getEclipseStrideTiaWarpConfig = async (
-  _routerConfig: ChainMap<RouterConfig>,
+  routerConfig: ChainMap<RouterConfig>,
 ): Promise<ChainMap<TokenRouterConfig>> => {
-  // @ts-ignore - foreignDeployment configs don't conform to the TokenRouterConfig
   const eclipsemainnet: TokenRouterConfig = {
+    ...routerConfig.eclipsemainnet,
     type: TokenType.synthetic,
     foreignDeployment: 'BpXHAiktwjx7fN6M9ST9wr6qKAsH27wZFhdHEhReJsR6',
     gas: 300_000,
   };
 
-  // @ts-ignore - foreignDeployment configs don't conform to the TokenRouterConfig
   const stride: TokenRouterConfig = {
+    ...routerConfig.stride,
     type: TokenType.collateral,
     foreignDeployment:
       'stride1pvtesu3ve7qn7ctll2x495mrqf2ysp6fws68grvcu6f7n2ajghgsh2jdj6',
+    token: 'stutia',
   };
 
   return {