+++ typescript/infra/scripts/check/check-owner-ica.ts
@@ -1,12 +1,20 @@
+import { ethers } from 'ethers';
+
 import { AccountConfig, InterchainAccount } from '@hyperlane-xyz/sdk';
-import { Address, eqAddress } from '@hyperlane-xyz/utils';
+import {
+  Address,
+  eqAddress,
+  isZeroish,
+  isZeroishAddress,
+} from '@hyperlane-xyz/utils';
 
 import { icas } from '../../config/environments/mainnet3/owners.js';
-import { getArgs as getEnvArgs } from '../agent-utils.js';
+import { isEthereumProtocolChain } from '../../src/utils/utils.js';
+import { getArgs as getEnvArgs, withChains } from '../agent-utils.js';
 import { getEnvironmentConfig, getHyperlaneCore } from '../core-utils.js';
 
 function getArgs() {
-  return getEnvArgs().option('ownerChain', {
+  return withChains(getEnvArgs()).option('ownerChain', {
     type: 'string',
     description: 'Origin chain where the Safe owner lives',
     default: 'ethereum',
@@ -14,7 +22,7 @@ function getArgs() {
 }
 
 async function main() {
-  const { environment, ownerChain } = await getArgs();
+  const { environment, ownerChain, chains } = await getArgs();
   const config = getEnvironmentConfig(environment);
   const multiProvider = await config.getMultiProvider();
 
@@ -29,17 +37,37 @@ async function main() {
   const { chainAddresses } = await getHyperlaneCore(environment, multiProvider);
   const ica = InterchainAccount.fromAddressesMap(chainAddresses, multiProvider);
 
+  const checkOwnerIcaChains = (
+    chains?.length ? chains : Object.keys(icas)
+  ).filter(isEthereumProtocolChain);
+
   const ownerConfig: AccountConfig = {
     origin: ownerChain,
     owner: owner,
   };
+  const ownerChainInterchainAccountRouter =
+    ica.contractsMap[ownerChain].interchainAccountRouter.address;
+
+  if (isZeroishAddress(ownerChainInterchainAccountRouter)) {
+    console.error(`Interchain account router address is zero`);
+    process.exit(1);
+  }
 
   const mismatchedResults: Record<
     string,
     { Expected: Address; Actual: Address }
   > = {};
-  for (const [chain, expectedAddress] of Object.entries(icas)) {
-    const actualAccount = await ica.getAccount(chain, ownerConfig);
+  for (const chain of checkOwnerIcaChains) {
+    const expectedAddress = icas[chain as keyof typeof icas];
+    if (!expectedAddress) {
+      console.error(`No expected address found for ${chain}`);
+      continue;
+    }
+    const actualAccount = await ica.getAccount(
+      chain,
+      ownerConfig,
+      ownerChainInterchainAccountRouter,
+    );
     if (!eqAddress(expectedAddress, actualAccount)) {
       mismatchedResults[chain] = {
         Expected: expectedAddress,
@@ -55,6 +83,7 @@ async function main() {
   } else {
     console.log('✅ All ICAs match the expected addresses.');
   }
+  process.exit(0);
 }
 
 main().catch((err) => {