+++ typescript/infra/scripts/warp-routes/deploy-warp-monitor.ts
@@ -1,32 +1,58 @@
+import { checkbox } from '@inquirer/prompts';
 import yargs from 'yargs';
 
 import { Contexts } from '../../config/contexts.js';
+import { WarpRouteIds } from '../../config/environments/mainnet3/warp/warpIds.js';
 import { HelmCommand } from '../../src/utils/helm.js';
 import { WarpRouteMonitorHelmManager } from '../../src/warp/helm.js';
-import { assertCorrectKubeContext, getAgentConfig } from '../agent-utils.js';
+import {
+  assertCorrectKubeContext,
+  getAgentConfig,
+  getArgs,
+  withWarpRouteId,
+} from '../agent-utils.js';
 import { getEnvironmentConfig } from '../core-utils.js';
 
 async function main() {
-  const { filePath } = await yargs(process.argv.slice(2))
-    .alias('f', 'filePath')
-    .describe(
-      'filePath',
-      'indicate the filepath to the warp route yaml file relative to the monorepo root',
-    )
-    .demandOption('filePath')
-    .string('filePath')
-    .parse();
+  const { environment, warpRouteId } = await withWarpRouteId(getArgs()).argv;
 
-  const environment = 'mainnet3';
+  let warpRouteIds;
+  if (warpRouteId) {
+    warpRouteIds = [warpRouteId];
+  } else {
+    warpRouteIds = await getWarpRouteIdsInteractive();
+  }
+
   await assertCorrectKubeContext(getEnvironmentConfig(environment));
   const agentConfig = getAgentConfig(Contexts.Hyperlane, environment);
 
-  const helmManager = new WarpRouteMonitorHelmManager(
-    filePath,
-    environment,
-    agentConfig.environmentChainNames,
-  );
-  await helmManager.runHelmCommand(HelmCommand.InstallOrUpgrade);
+  const deployWarpMonitor = async (warpRouteId: string) => {
+    const helmManager = new WarpRouteMonitorHelmManager(
+      warpRouteId,
+      environment,
+      agentConfig.environmentChainNames,
+    );
+    await helmManager.runHelmCommand(HelmCommand.InstallOrUpgrade);
+  };
+
+  for (const id of warpRouteIds) {
+    console.log(`Deploying Warp Monitor for Warp Route ID: ${id}`);
+    await deployWarpMonitor(id);
+  }
+}
+
+async function getWarpRouteIdsInteractive() {
+  const choices = Object.values(WarpRouteIds).map((id) => ({
+    value: id,
+  }));
+
+  const selection = await checkbox({
+    message: 'Select Warp Route IDs to deploy',
+    choices,
+    pageSize: 30,
+  });
+
+  return selection;
 }
 
 main()