+++ typescript/infra/test/govern.hardhat-test.ts
@@ -1,6 +1,6 @@
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers.js';
 import { expect } from 'chai';
-import { BigNumber, utils } from 'ethers';
+import { BigNumber } from 'ethers';
 import hre from 'hardhat';
 
 import {
@@ -27,6 +27,7 @@ import {
   TestChainName,
   TestCoreApp,
   TestCoreDeployer,
+  randomAddress,
 } from '@hyperlane-xyz/sdk';
 import { Address, CallData, eqAddress } from '@hyperlane-xyz/utils';
 
@@ -34,9 +35,6 @@ import {
   AnnotatedCallData,
   HyperlaneAppGovernor,
 } from '../src/govern/HyperlaneAppGovernor.js';
-
-// TODO de-dupe with test-utils after migrating this file to the SDK
-const randomAddress = () => utils.hexlify(utils.randomBytes(20));
 
 export class TestApp extends HyperlaneApp<{}> {}