+++ typescript/sdk/src/providers/transactions/submitter/schemas.ts
@@ -5,12 +5,13 @@ import {
   EV5GnosisSafeTxBuilderPropsSchema,
   EV5GnosisSafeTxSubmitterPropsSchema,
   EV5ImpersonatedAccountTxSubmitterPropsSchema,
+  EV5JsonRpcTxSubmitterPropsSchema,
 } from './ethersV5/schemas.js';
 
 export const SubmitterMetadataSchema = z.discriminatedUnion('type', [
   z.object({
     type: z.literal(TxSubmitterType.JSON_RPC),
-    privateKey: z.string().optional(),
+    ...EV5JsonRpcTxSubmitterPropsSchema.shape,
   }),
   z.object({
     type: z.literal(TxSubmitterType.IMPERSONATED_ACCOUNT),