+++ typescript/sdk/src/token/adapters/SealevelTokenAdapter.ts
@@ -113,6 +113,11 @@ export class SealevelNativeTokenAdapter
       }),
     );
   }
+
+  async getTotalSupply(): Promise<bigint | undefined> {
+    // Not implemented.
+    return undefined;
+  }
 }
 
 // Interacts with SPL token programs
@@ -190,6 +195,13 @@ export class SealevelTokenAdapter
       this.getTokenProgramId(),
     );
   }
+
+  async getTotalSupply(): Promise<bigint | undefined> {
+    const response = await this.getProvider().getTokenSupply(
+      this.tokenMintPubKey,
+    );
+    return BigInt(response.value.amount);
+  }
 }
 
 interface HypTokenAddresses {
@@ -264,6 +276,11 @@ export abstract class SealevelHypTokenAdapter
       domain,
       address: pubKey.toBuffer(),
     }));
+  }
+
+  // Intended to be overridden by subclasses
+  async getBridgedSupply(): Promise<bigint | undefined> {
+    return undefined;
   }
 
   async quoteTransferRemoteGas(
@@ -581,6 +598,10 @@ export class SealevelHypNativeAdapter extends SealevelHypTokenAdapter {
     return this.wrappedNative.getBalance(owner);
   }
 
+  override async getBridgedSupply(): Promise<bigint> {
+    return this.getBalance(this.addresses.warpRouter);
+  }
+
   override async getMetadata(): Promise<TokenMetadata> {
     return this.wrappedNative.getMetadata();
   }
@@ -630,6 +651,10 @@ export class SealevelHypCollateralAdapter extends SealevelHypTokenAdapter {
     }
 
     return super.getBalance(owner);
+  }
+
+  override async getBridgedSupply(): Promise<bigint> {
+    return this.getBalance(this.addresses.warpRouter);
   }
 
   override getTransferInstructionKeyList(
@@ -695,6 +720,10 @@ export class SealevelHypSyntheticAdapter extends SealevelHypTokenAdapter {
       if (error.message?.includes(NON_EXISTENT_ACCOUNT_ERROR)) return 0n;
       throw error;
     }
+  }
+
+  override async getBridgedSupply(): Promise<bigint> {
+    return this.getTotalSupply();
   }
 
   async getTotalSupply(): Promise<bigint> {